update exception handler, move classes
