+++ src/main/java/com/forest/mytopmovies/entity/Movie.java
@@ -1,21 +1,14 @@
 package com.forest.mytopmovies.entity;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
-import javax.persistence.Table;
+import lombok.Builder;
+
+import javax.persistence.*;
 import java.util.Date;
 import java.util.List;
 
 @Entity
 @Table(name = "mtm_movies")
+@Builder
 public class Movie {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -39,8 +32,14 @@ public class Movie {
     @Column(name = "release_date")
     private Date releaseDate;
 
-    @Column(name = "genre_ids")
-    private String genreIds;
+    @ManyToMany(fetch = FetchType.LAZY,
+            cascade = {CascadeType.PERSIST, CascadeType.MERGE
+                    , CascadeType.DETACH, CascadeType.REFRESH})
+    @JoinTable(
+            name = "mtm_movie_genre",
+            joinColumns = @JoinColumn(name = "movie_id", referencedColumnName = "id"),
+            inverseJoinColumns = @JoinColumn(name = "genre_id ", referencedColumnName = "id"))
+    private List<Genres> genres;
 
     @Column(name = "tmdb_id")
     private int tmdbId;
@@ -49,7 +48,7 @@ public class Movie {
             cascade = {CascadeType.PERSIST, CascadeType.MERGE
                     , CascadeType.DETACH, CascadeType.REFRESH})
     @JoinTable(
-            name = "mtm_movie_movielist",
+            name = "mtm_movie_movie_list",
             joinColumns = @JoinColumn(name = "movie_id", referencedColumnName = "id"),
             inverseJoinColumns = @JoinColumn(name = "movielist_id ", referencedColumnName = "id"))
     private List<MovieList> movieLists;
@@ -110,12 +109,12 @@ public class Movie {
         this.releaseDate = releaseDate;
     }
 
-    public String getGenreIds() {
-        return genreIds;
+    public List<Genres> getGenres() {
+        return genres;
     }
 
-    public void setGenreIds(String genreIds) {
-        this.genreIds = genreIds;
+    public void setGenres(List<Genres> genres) {
+        this.genres = genres;
     }
 
     public int getTmdbId() {
